python-api/requirement.txt
# Main Python Server
fastapi
uvicorn[standard]
python-multipart

# Question Generation
transformers
torch
sentence-transformers
PyMuPDF
pytesseract
pdf2image
# You may need to install Poppler separately for pdf2image to work

# Summarizer
python-pptx